Conflicts:
	lib/Pipeline/RNASeq.pm

+++ lib/Alignment/STARFeatureCount.pm
@@ -21,6 +21,7 @@ sub new {
   my $self = $class->SUPER::new();
   $self->{_name}   = __PACKAGE__;
   $self->{_suffix} = "_sf";
+  $self->{_use_tmp_folder} = 1;
   bless $self, $class;
   return $self;
 }
@@ -99,11 +100,6 @@ sub perform {
 
   for my $sample_name ( sort keys %fqFiles ) {
     my @sample_files = @{ $fqFiles{$sample_name} };
-    my $sample_file_1 = $sample_files[0];
-
-    my $uncompress = ( $sample_file_1 =~ /.gz$/ ) ? " --readFilesCommand zcat" : "";
-
-    my $samples = join( " ", @sample_files );
 
     my $pbs_file = $self->get_pbs_filename( $pbs_dir, $sample_name );
     my $pbs_name = basename($pbs_file);
@@ -125,7 +121,14 @@ sub perform {
 
     my $chromosome_grep_command = $output_sort_by_coordinate ? getChromosomeFilterCommand( $final_bam, $chromosome_grep_pattern ) : "";
 
+    my $localized_files = [];
+    @sample_files = @{$self->localize_files_in_tmp_folder($pbs, \@sample_files, $localized_files)};
+    my $samples = join( " ", @sample_files );
+    my $sample_file_1 = $sample_files[0];
+    my $uncompress = ( $sample_file_1 =~ /.gz$/ ) ? " --readFilesCommand zcat" : "";
+
     print $pbs "
+
 if [[ ! -s $final_bam && -s $sample_file_1 ]]; then
   echo performing star ...
   $star $option --outSAMattrRGline $rgline --runThreadN $thread --genomeDir $star_index --readFilesIn $samples $uncompress --outFileNamePrefix ${sample_name}_ $output_format
@@ -145,8 +148,9 @@ fi
   if ($output_sort_by_coordinate) {
     print $pbs "
 if [ ! -s ${final_bam} ]; then
-  sambamba sort -m $sort_memory -t $thread -o $final_bam $unsorted
-  sambamba index $final_bam
+  samtools sort -m $sort_memory -T ${sample_name} -t $thread -o $final_bam $unsorted
+  samtools index $final_bam
+  samtools idxstats $final_bam > ${final_bam}.chromosome.count
 fi  
 ";
   }
@@ -170,7 +174,7 @@ fi
 
     if ( !$output_unsorted ) {
       print $pbs "
-if [ -s $final_file ]; then
+if [[ -s $final_file && -s $bam_stat ]]; then
   rm $unsorted 
 fi
 ";
@@ -185,8 +189,15 @@ fi
 ";
     }
 
+    $self->clean_temp_files($pbs, $localized_files);
+
     $self->close_pbs( $pbs, $pbs_file );
-    print $sh "\$MYCMD ./$pbs_name \n";
+
+    print $sh "
+if [[ ! -s $result_dir/$final_file ]]; then
+  \$MYCMD ./$pbs_name 
+fi
+";
   }
   print $sh "exit 0\n";
   close $sh;
@@ -222,15 +233,16 @@ sub result {
     push( @result_files, "$cur_dir/${sample_name}.count.summary" );
     push( @result_files, "$cur_dir/${sample_name}_Log.final.out" );
     push( @result_files, "$cur_dir/${sample_name}.bamstat" );
+    push( @result_files, "$cur_dir/${sample_name}.count" );
     if (!$delete_star_featureCount_bam) {
       if ($output_sort_by_coordinate) {
         push( @result_files, "$cur_dir/${sample_name}_Aligned.sortedByCoord.out.bam" );
+        push( @result_files, "$cur_dir/${sample_name}_Aligned.sortedByCoord.out.bam.chromosome.count" );
       }
       if ($output_unsorted) {
         push( @result_files, "$cur_dir/${sample_name}_Aligned.out.bam" );
       }
     }
-    push( @result_files, "$cur_dir/${sample_name}.count" );
     push( @result_files, "$cur_dir/${sample_name}.count.star.version" );
     push( @result_files, "$cur_dir/${sample_name}.count.featureCounts.version" );
     $result->{$sample_name} = filter_array( \@result_files, $pattern );
